refactor(analysis): migrate build_snapshot_diff from agents to analysis layer
Migrate `build_snapshot_diff` function from `vibe3.agents.review_pipeline_helpers`
to `vibe3.analysis.snapshot_service`, eliminating redundant wrapper layer.

Changes:
- Add `build_snapshot_diff` to snapshot_service.py with local import to avoid circular dependency
- Register function in analysis/__init__.py (lazy imports + __all__)
- Update roles/review.py to import from vibe3.analysis instead of vibe3.agents
- Remove build_snapshot_diff from agents/__init__.py exports and docstring
- Delete review_pipeline_helpers.py module (single-function module no longer needed)
- Migrate tests to tests/vibe3/analysis/test_snapshot_diff_builder.py
- Update agents/README.md to reflect removed module

Fixes #2827
